Add centered BB_Ball + created Triangle and Mesh class

+++ HelloWorld/RayTriangleIntersect.cpp
@@ -22,9 +22,9 @@
 namespace Intersection {
 
 	/* the original jgt code */
-	int intersect_triangle(double orig[3], double dir[3],
-		double vert0[3], double vert1[3], double vert2[3],
-		double* t, double* u, double* v)
+	int intersect_triangle(float orig[3], float dir[3],
+		float vert0[3], float vert1[3], float vert2[3],
+		float* t, float* u, float* v)
 	{
 		double edge1[3], edge2[3], tvec[3], pvec[3], qvec[3];
 		double det, inv_det;
